Move `replaceSectionTypes` logic into CurriculumModel createFromJson function
